% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataframe.R
\name{create_dataframe}
\alias{create_dataframe}
\title{Create dataframe}
\usage{
create_dataframe(num_participants, n, min_num_correct, num_rep, angles, p_t0)
}
\arguments{
\item{num_participants}{Number of participants}

\item{n}{Number of target items (not implemented yet, n = 4)}

\item{min_num_correct}{Minimum number of correct tracked objects}

\item{num_rep}{Number of repetitions / simulations}

\item{angles}{Reflection angles (not implemented yet)}

\item{p_t0}{Probability of selecting target 1 (singleton)}
}
\description{
This function creates a dataframe with the simulated MOT performance of n targets.
}
\author{
Markus Huff
}

